refactoring min cost flow solver in Payment Class
