+++ pickhardtpayments/Payment.py
@@ -1,6 +1,11 @@
 import time
 from typing import List
 
+from ortools.graph import pywrapgraph
+import networkx as nx
+
+from .UncertaintyNetwork import UncertaintyNetwork
+from .OracleLightningNetwork import OracleLightningNetwork
 from .Attempt import Attempt, AttemptStatus
 
 
@@ -21,7 +26,8 @@ class Payment:
     :type total_amount: class:`int`
     """
 
-    def __init__(self, sender, receiver, total_amount: int = 1):
+    def __init__(self, uncertainty_network: UncertaintyNetwork, oracle_network: OracleLightningNetwork, sender,
+                 receiver, total_amount: int = 1):
         """Constructor method
         """
         self._successful = False
@@ -33,6 +39,9 @@ class Payment:
         self._receiver = receiver
         self._total_amount = total_amount
         self._attempts = list()
+        self._uncertainty_network = uncertainty_network
+        self._oracle_network = oracle_network
+        self._prepare_integer_indices_for_nodes()
 
     def __str__(self):
         return "Payment with {} attempts to deliver {} sats from {} to {}".format(len(self._attempts),
@@ -149,7 +158,7 @@ class Payment:
         :return: fee in ppm for successful delivery and settlement of payment
         :rtype: float
         """
-        return self.fee * 1000 / self.total_amount
+        return self._fee * 1000 / self.total_amount
 
     def filter_attempts(self, flag: AttemptStatus) -> List[Attempt]:
         """Returns all onions with the given state.
@@ -181,3 +190,179 @@ class Payment:
         :type: bool
         """
         self._successful = value
+
+    def _prepare_integer_indices_for_nodes(self):
+        """
+        necessary for the OR-lib by google and the min cost flow solver
+
+        let's initialize the look-up tables for node_ids to integers from [0,...,#number of nodes]
+        this is necessary because of the API of the Google Operations Research min cost flow solver
+        """
+        self._mcf_id = {}
+        self._node_key = {}
+        for k, node_id in enumerate(self._uncertainty_network.network.nodes()):
+            self._mcf_id[node_id] = k
+            self._node_key[k] = node_id
+
+    def generate_candidate_paths(self, mu: int, base: int):
+        """
+        computes the optimal payment split to deliver `amt` from `src` to `dest` and updates our belief about the
+        liquidity
+
+        This is one step within the payment loop.
+
+        Returns the residual amount of the `amt` that could ne be delivered and the paid fees
+        (on a per channel base not including fees for downstream fees) for the delivered amount
+
+        the function also prints some results on statistics about the paths of the flow to stdout.
+        """
+        # initialisation of List of Attempts for this round.
+        attempts_in_round = List[Attempt]
+
+        # First we prepare the min cost flow by getting arcs from the uncertainty network
+        self._prepare_mcf_solver(mu, base)
+        self._start_time = time.time()
+        # print("solving mcf...")
+        status = self._min_cost_flow.Solve()
+
+        if status != self._min_cost_flow.OPTIMAL:
+            print('There was an issue with the min cost flow input.')
+            print(f'Status: {status}')
+            exit(1)
+
+        attempts_in_round = self._dissect_flow_to_paths(self._sender, self._receiver)
+        self._end_time = time.time()
+        self.add_attempts(attempts_in_round)
+        return 0
+
+    def _prepare_mcf_solver(self, mu: int, base_fee: int):
+        """
+        computes the uncertainty network given our prior belief and prepares the min cost flow solver
+
+        This function can define a value for mu to control how heavily we combine the uncertainty cost and fees Also
+        the function supports only taking channels into account that don't charge a base_fee higher or equal to `base`
+
+        returns the instantiated min_cost_flow object from the Google OR-lib that contains the piecewise linearized
+        problem
+        """
+        self._min_cost_flow = pywrapgraph.SimpleMinCostFlow()
+        self._arc_to_channel = {}
+
+        for s, d, channel in self._uncertainty_network.network.edges(data="channel"):
+            # ignore channels with too large base fee
+            if channel.base_fee > base_fee:
+                continue
+            # FIXME: Remove Magic Number for pruning
+            # Prune channels away that have too low success probability! This is a huge runtime boost
+            # However the pruning would be much better to work on quantiles of normalized cost
+            # So as soon as we have better Scaling, Centralization and feature engineering we can
+            # probably have a more focused pruning
+            if self._uncertainty_network.prune and channel.success_probability(250_000) < 0.9:
+                continue
+            cnt = 0
+            # QUANTIZATION):
+            for capacity, cost in channel.get_piecewise_linearized_costs(mu=mu):
+                index = self._min_cost_flow.AddArcWithCapacityAndUnitCost(self._mcf_id[s],
+                                                                          self._mcf_id[d],
+                                                                          capacity,
+                                                                          cost)
+                self._arc_to_channel[index] = (s, d, channel, 0)
+                if self._uncertainty_network.prune and cnt > 1:
+                    break
+                cnt += 1
+
+        # Add node supply to 0 for all nodes
+        for i in self._uncertainty_network.network.nodes():
+            self._min_cost_flow.SetNodeSupply(self._mcf_id[i], 0)
+
+        # add amount to sending node
+        self._min_cost_flow.SetNodeSupply(
+            self._mcf_id[self._sender], int(self._total_amount))  # /QUANTIZATION))
+
+        # add -amount to recipient nods
+        self._min_cost_flow.SetNodeSupply(
+            self._mcf_id[self._receiver], -int(self._total_amount))  # /QUANTIZATION))
+
+    def _make_channel_path(self, G: nx.MultiDiGraph, path: List[str]) -> object:
+        """
+        network x returns a path as a list of node_ids. However, we need a list of `UncertaintyChannels`
+        Since the graph has parallel edges it is quite some work to get the actual channels that the
+        min cost flow solver produced
+        """
+        channel_path = []
+        bottleneck = 2 ** 63
+        for src, dest in self._next_hop(path):
+            w = 2 ** 63
+            c = None
+            flow = 0
+            for sid in G[src][dest].keys():
+                if G[src][dest][sid]["weight"] < w:
+                    w = G[src][dest][sid]["weight"]
+                    c = G[src][dest][sid]["channel"]
+                    flow = G[src][dest][sid]["flow"]
+            channel_path.append(c)
+
+            if flow < bottleneck:
+                bottleneck = flow
+
+        return channel_path, bottleneck
+
+    def _dissect_flow_to_paths(self, s, d):
+        """
+        A standard algorithm to dissect a flow into several paths.
+
+
+        FIXME: Note that this dissection while accurate is probably not optimal in practice.
+        As noted in our Probabilistic payment delivery paper the payment process is a bernoulli trial
+        and I assume it makes sense to dissect the flow into paths of similar likelihood to make most
+        progress but this is a mere conjecture at this point. I expect quite a bit of research will be
+        necessary to resolve this issue.
+        """
+        # first collect all linearized edges which are assigned a non-zero flow put them into a networkx graph
+        G = nx.MultiDiGraph()
+        for i in range(self._min_cost_flow.NumArcs()):
+            flow = self._min_cost_flow.Flow(i)  # *QUANTIZATION
+            if flow == 0:
+                continue
+
+            src, dest, channel, _ = self._arc_to_channel[i]
+            if G.has_edge(src, dest):
+                if channel.short_channel_id in G[src][dest]:
+                    G[src][dest][channel.short_channel_id]["flow"] += flow
+            else:
+                # FIXME: cost is not reflecting exactly the piecewise linearization
+                # Probably not such a big issue as we just dissect flow
+                G.add_edge(src, dest, key=channel.short_channel_id, flow=flow,
+                           channel=channel, weight=channel.combined_linearized_unit_cost())
+        used_flow = 1
+        attempts = []
+
+        # allocate flow to shortest / cheapest paths from src to dest as long as this is possible
+        # decrease flow along those edges. This is a standard mechanism to dissect a flow into paths
+        while used_flow > 0:
+            try:
+                path = nx.shortest_path(G, s, d)
+            except:
+                break
+            channel_path, used_flow = self._make_channel_path(G, path)
+            attempts.append(Attempt(channel_path, used_flow))
+
+            # reduce the flow from the selected path
+            for pos, hop in enumerate(self._next_hop(path)):
+                src, dest = hop
+                channel = channel_path[pos]
+                G[src][dest][channel.short_channel_id]["flow"] -= used_flow
+                if G[src][dest][channel.short_channel_id]["flow"] == 0:
+                    G.remove_edge(src, dest, key=channel.short_channel_id)
+        return attempts
+
+    def _next_hop(self, path):
+        """
+        generator to iterate through edges indexed by node id of paths
+
+        The path is a list of node ids. Each call returns a tuple src, dest of an edge in the path
+        """
+        for i in range(1, len(path)):
+            src = path[i - 1]
+            dest = path[i]
+            yield src, dest
